@@ -2,7 +2,7 @@ name: webview_flutter_example
 description: Demonstrates how to use the webview_flutter plugin.
 
 environment:
-  sdk: ">=2.0.0-dev.68.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
 
 dependencies:
   flutter:
