@@ -2,6 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+// @dart = 2.9
+
 import 'package:android_intent/flag.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -11,7 +13,8 @@ import 'package:platform/platform.dart';
 
 void main() {
   AndroidIntent androidIntent;
-  late MockMethodChannel mockChannel;
+  MockMethodChannel mockChannel;
+
   setUp(() {
     mockChannel = MockMethodChannel();
     when(mockChannel.invokeMethod<bool>('canResolveActivity', any))
