@@ -2,6 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+// @dart = 2.9
+
 import 'dart:io';
 import 'dart:ui';
 
@@ -15,7 +17,7 @@ import 'package:flutter/services.dart';
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
 
-  late MockMethodChannel mockChannel;
+  MockMethodChannel mockChannel;
 
   setUp(() {
     mockChannel = MockMethodChannel();
